Solved

+++ Step14/2_2748.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 using namespace std;
 
-/* for문이 재귀함수보다 빠르다 */
+/* int형으로 표현할 수 있는 수를 넘는다 -> long long형 */
 int main() {
 	int n;
 	cin >> n;
 
-	int fibonacci[n+1];
+	long long fibonacci[n+1];
 	fibonacci[0] = 0;
 	fibonacci[1] = 1;
 	for(int i = 2; i < n + 1; i++) {
